moved signing code out of main
